Fix MSVC warnings about integer overflow

+++ core/base/bottleneckDistance/BottleneckDistanceImpl.h
@@ -1,7 +1,7 @@
 #ifndef _BOTTLENECKDISTANCEIMPL_H
 #define _BOTTLENECKDISTANCEIMPL_H
 
-constexpr unsigned int str2int(const char *str, int h = 0) {
+constexpr unsigned long long str2int(const char *str, int h = 0) {
   return !str[h] ? 5381 : (str2int(str, h + 1) * 33) ^ str[h];
 }
