+++ standalone/LDistance/gui/main.cpp
@@ -12,7 +12,7 @@ using namespace ttk;
 
 vtkUserInterface<ttkLDistance> program;
 
-constexpr unsigned int str2int(const char *str, int h = 0) {
+constexpr unsigned long long str2int(const char *str, int h = 0) {
   return !str[h] ? 5381 : (str2int(str, h + 1) * 33) ^ str[h];
 }
